Tidy up and seperate
