+++ packages/istanbul-report-modern-html/src/filterButtons.js
@@ -1,0 +1,46 @@
+import * as React from 'react';
+
+function FilterButton({ children, filter, activeFilters, setFilters }) {
+    return (
+        <button
+            class={'togglebutton ' + (activeFilters[filter] ? 'enabled' : '')}
+            onClick={() =>
+                setFilters({
+                    ...activeFilters,
+                    [filter]: !activeFilters[filter]
+                })
+            }
+        >
+            {children}
+        </button>
+    );
+}
+
+export default function FilterButtons({ activeFilters, setFilters }) {
+    return (
+        <div class="buttongroup">
+            <label>Coverage:</label>
+            <FilterButton
+                filter="low"
+                activeFilters={activeFilters}
+                setFilters={setFilters}
+            >
+                Low
+            </FilterButton>
+            <FilterButton
+                filter="medium"
+                activeFilters={activeFilters}
+                setFilters={setFilters}
+            >
+                Medium
+            </FilterButton>
+            <FilterButton
+                filter="high"
+                activeFilters={activeFilters}
+                setFilters={setFilters}
+            >
+                High
+            </FilterButton>
+        </div>
+    );
+}
