+++ packages/istanbul-report-modern-html/src/index.js
@@ -4,144 +4,11 @@ import * as ReactDOM from 'react-dom';
 import SummaryTableHeader from './summaryTableHeader';
 import SummaryTableLine from './summaryTableLine';
 import SummaryHeader from './summaryHeader';
+import getChildData from './getChildData';
+import SummarizerButtons from './summarizerButtons';
+import FilterButtons from './filterButtons';
 
 const sourceData = window.data;
-
-function addPath(node, parentPath) {
-    if (!parentPath) {
-        return node;
-    }
-    return { ...node, file: parentPath + '/' + node.file };
-}
-
-function flatten(nodes, parentPath) {
-    let children = [];
-    for (var i = 0; i < nodes.length; i++) {
-        let child = nodes[i];
-        if (child.children) {
-            children = [
-                ...children.map(child => addPath(child, parentPath)),
-                ...flatten(
-                    child.children,
-                    (!parentPath ? '' : '/' + parentPath) + child.file
-                )
-            ];
-        } else {
-            children.push(addPath(child, parentPath));
-        }
-    }
-    return children;
-}
-
-function sort(childData, activeSort) {
-    const top = activeSort.order === 'asc' ? 1 : -1;
-    const bottom = activeSort.order === 'asc' ? -1 : 1;
-    childData.sort((a, b) => {
-        let valueA;
-        let valueB;
-        if (activeSort.sortKey === 'file') {
-            // reverse to match original report ordering
-            valueB = a.file;
-            valueA = b.file;
-        } else {
-            const [metricType, valueType] = activeSort.sortKey.split('.');
-            valueA = a.metrics[metricType][valueType];
-            valueB = b.metrics[metricType][valueType];
-        }
-
-        if (valueA === valueB) {
-            return 0;
-        }
-        return valueA < valueB ? top : bottom;
-    });
-
-    for (let i = 0; i < childData.length; i++) {
-        let child = childData[i];
-        if (child.children) {
-            childData[i] = {
-                ...child,
-                children: sort(child.children, activeSort)
-            };
-        }
-    }
-    return childData;
-}
-
-function filter(nodes, activeFilters) {
-    let children = [];
-    for (var i = 0; i < nodes.length; i++) {
-        let child = nodes[i];
-        if (child.children) {
-            const newSubChildren = filter(child.children, activeFilters);
-            if (newSubChildren.length) {
-                child = { ...child, children: newSubChildren };
-                children.push(child);
-            }
-        } else {
-            if (
-                activeFilters[child.metrics.statements.classForPercent] ||
-                activeFilters[child.metrics.branches.classForPercent] ||
-                activeFilters[child.metrics.functions.classForPercent] ||
-                activeFilters[child.metrics.lines.classForPercent]
-            ) {
-                children.push(child);
-            }
-        }
-    }
-    return children;
-}
-
-function getChildData(activeSort, summarizerType, activeFilters) {
-    let childData;
-
-    if (summarizerType === 'flat') {
-        childData = flatten(sourceData['package'].children.slice(0));
-    } else {
-        childData = sourceData[summarizerType].children;
-    }
-
-    childData = filter(childData, activeFilters);
-
-    if (activeSort) {
-        childData = sort(childData, activeSort);
-    }
-    return childData;
-}
-
-function SummarizerButton({
-    summarizerType,
-    activeSummarizerType,
-    setSummarizerType,
-    children
-}) {
-    return (
-        <button
-            onClick={() => setSummarizerType(summarizerType)}
-            class={
-                'togglebutton ' +
-                (summarizerType === activeSummarizerType ? 'enabled' : '')
-            }
-        >
-            {children}
-        </button>
-    );
-}
-
-function FilterButton({ children, filter, activeFilters, setFilters }) {
-    return (
-        <button
-            class={'togglebutton ' + (activeFilters[filter] ? 'enabled' : '')}
-            onClick={() =>
-                setFilters({
-                    ...activeFilters,
-                    [filter]: !activeFilters[filter]
-                })
-            }
-        >
-            {children}
-        </button>
-    );
-}
 
 function App() {
     const [activeSort, setSort] = React.useState({
@@ -155,7 +22,8 @@ function App() {
         high: true
     });
     const childData = React.useMemo(
-        () => getChildData(activeSort, summarizerType, activeFilters),
+        () =>
+            getChildData(sourceData, activeSort, summarizerType, activeFilters),
         [activeSort, summarizerType, activeFilters]
     );
     const overallMetrics = sourceData.package.metrics;
@@ -165,54 +33,14 @@ function App() {
             <div class="wrapper">
                 <SummaryHeader metrics={overallMetrics} />
                 <div class="pad1">
-                    <div class="buttongroup">
-                        <label>Summarizer:</label>
-                        <SummarizerButton
-                            setSummarizerType={setSummarizerType}
-                            summarizerType="package"
-                            activeSummarizerType={summarizerType}
-                        >
-                            Package
-                        </SummarizerButton>
-                        <SummarizerButton
-                            setSummarizerType={setSummarizerType}
-                            summarizerType="nested"
-                            activeSummarizerType={summarizerType}
-                        >
-                            Nested
-                        </SummarizerButton>
-                        <SummarizerButton
-                            setSummarizerType={setSummarizerType}
-                            summarizerType="flat"
-                            activeSummarizerType={summarizerType}
-                        >
-                            Flat
-                        </SummarizerButton>
-                    </div>
-                    <div class="buttongroup">
-                        <label>Coverage:</label>
-                        <FilterButton
-                            filter="low"
-                            activeFilters={activeFilters}
-                            setFilters={setFilters}
-                        >
-                            Low
-                        </FilterButton>
-                        <FilterButton
-                            filter="medium"
-                            activeFilters={activeFilters}
-                            setFilters={setFilters}
-                        >
-                            Medium
-                        </FilterButton>
-                        <FilterButton
-                            filter="high"
-                            activeFilters={activeFilters}
-                            setFilters={setFilters}
-                        >
-                            High
-                        </FilterButton>
-                    </div>
+                    <SummarizerButtons
+                        setSummarizerType={setSummarizerType}
+                        summarizerType={summarizerType}
+                    />
+                    <FilterButtons
+                        activeFilters={activeFilters}
+                        setFilters={setFilters}
+                    />
                 </div>
                 <div class="pad1">
                     <table class="coverage-summary">
